delete account cookies

+++ jero-frontend/components/Profile/ManageProfile/ManageProfile.tsx
@@ -47,11 +47,15 @@ const ManageProfile = (props : UpdateFields) => {
 
     const handleDelete = async (e : any) => {
         e.preventDefault();
+        if(!confirm("delete")){
+            return;
+        }
         try {
             const response = await axios.delete(`${baseApi}/auth/delete`,
                 { withCredentials: true}
             );
             console.log("hi" + response.data);
+            alert("account deleted")
         } catch (error : any) {
             console.log(error)
             //onsole.log('Login failed:', error.response ? error.response.data : error.message);
